requirements.txt
Django>=1.9.0
jsonfield>=2.0.2
requests>=2.18.4
django-jsoneditor>=0.0.12